+++ frontend/src/tests/lib/pages/SnsWallet.spec.ts
@@ -16,16 +16,11 @@ import { mockSnsMainAccount } from "$tests/mocks/sns-accounts.mock";
 import { snsResponseFor } from "$tests/mocks/sns-response.mock";
 import { Principal } from "@dfinity/principal";
 import { SnsSwapLifecycle } from "@dfinity/sns";
-import {
-  fireEvent,
-  render,
-  waitFor,
-  type RenderResult,
-} from "@testing-library/svelte";
-import type { SvelteComponent } from "svelte";
+import { fireEvent, render, waitFor } from "@testing-library/svelte";
 import { get } from "svelte/store";
 import en from "../../mocks/i18n.mock";
 import { waitModalIntroEnd } from "../../mocks/modal.mock";
+import { testAccountsModal } from "../../utils/accounts.test-utils";
 import AccountsTest from "./AccountsTest.svelte";
 
 jest.mock("$lib/services/sns-accounts.services", () => {
@@ -128,26 +123,6 @@ describe("SnsWallet", () => {
       );
     });
 
-    const testModal = async ({
-      result,
-      testId,
-    }: {
-      result: RenderResult<SvelteComponent>;
-      testId: string;
-    }) => {
-      const { container, getByTestId } = result;
-
-      await waitFor(expect(getByTestId(testId)).not.toBeNull);
-
-      const button = getByTestId(testId) as HTMLButtonElement;
-
-      await fireEvent.click(button);
-
-      await waitFor(() =>
-        expect(container.querySelector("div.modal")).not.toBeNull()
-      );
-    };
-
     it("should open new transaction modal", async () => {
       const result = render(SnsWallet, props);
 
@@ -157,7 +132,7 @@ describe("SnsWallet", () => {
         expect(queryByTestId("open-new-sns-transaction")).toBeInTheDocument()
       );
 
-      await testModal({ result, testId: "open-new-sns-transaction" });
+      await testAccountsModal({ result, testId: "open-new-sns-transaction" });
 
       await waitFor(() => {
         expect(getByTestId("transaction-step-1")).toBeInTheDocument();
@@ -172,7 +147,7 @@ describe("SnsWallet", () => {
     it("should open receive modal", async () => {
       const result = render(AccountsTest, { props: modalProps });
 
-      await testModal({ result, testId: "receive-sns" });
+      await testAccountsModal({ result, testId: "receive-sns" });
 
       const { getByTestId } = result;
 
@@ -187,7 +162,7 @@ describe("SnsWallet", () => {
 
       const result = render(AccountsTest, { props: modalProps });
 
-      await testModal({ result, testId: "receive-sns" });
+      await testAccountsModal({ result, testId: "receive-sns" });
 
       const { getByTestId, container } = result;
 
@@ -207,7 +182,7 @@ describe("SnsWallet", () => {
     it("should display receive modal information", async () => {
       const result = render(AccountsTest, { props: modalProps });
 
-      await testModal({ result, testId: "receive-sns" });
+      await testAccountsModal({ result, testId: "receive-sns" });
 
       const { getByText } = result;
